Extract BaseTagsAndCategoriesViewModel and use localized error messages
